Move PageLimitController and LimitAndPageControls to browse/components directory. TODO: Move over browse and search pages and any reusable components.
