+++ src/encoded/static/components/browse/components/index.js
@@ -1,0 +1,4 @@
+'use strict';
+
+export { LimitAndPageControls } from './LimitAndPageControls';
+export { PageLimitSortController } from './PageLimitSortController';
